requirements.txt

accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
astar==0.93
astor==0.8.1
attrs==25.3.0
blake3==1.0.4
cachelib==0.10.2
cachetools==5.5.2
certifi==2025.1.31
cloudpickle==3.1.1
compressed-tensors==0.9.2
contourpy==1.0.7
cupy-cuda12x==13.4.0
cycler==0.11.0
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
Flask==2.2.3
flask-google-cloud-logger==0.2.1
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
fonttools==4.39.0
frozenlist==1.5.0
fsspec==2025.3.0
gguf==0.10.0
google-cloud-logger==0.2.1
greenlet==2.0.2
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
itsdangerous==2.1.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.4
langchain
langchain-openai
langgraph
langsmith
lark==1.2.2
# llm
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
mistral_common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama
openai==1.67.0
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.1.0
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
pycountry==24.6.1
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.1.0
Pygments==2.19.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.23.1
safetensors==0.5.3
scipy==1.13.1
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.6
starlette==0.46.1
sympy==1.13.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.8.1
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==2.2.3
wheel==0.45.1
xformers==0.0.29.post2
xgrammar==0.1.16
yarl==1.18.3
zipp==3.21.0
